requirements.txt
anyio==4.12.0
apsw==3.51.1.0
apswutils==0.1.1
beautifulsoup4==4.14.3
certifi==2025.11.12
click==8.3.1
fastcore==1.8.17
fastlite==0.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
lxml==6.0.2
mistletoe==1.5.1
monsterui==1.0.34
oauthlib==3.3.1
packaging==25.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-fasthtml==0.12.36
python-multipart==0.0.20
PyYAML==6.0.3
six==1.17.0
soupsieve==2.8
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
